requirements.txt
# Usage: pip install -r requirements.txt
# if no good network, use -i https://pypi.tuna.tsinghua.edu.cn/simple

torchmetrics==0.11.4
opencv-python==4.7.0.72
numpy==1.24.3
tqdm==4.65.0
Pillow==9.4.0
grad-cam==1.4.8
timm==0.9.16
tensorboard==2.16.2
prettytable==3.10.0